refactor(agents): 스킬 triggers 로직을 skill-triggers.ts 로 분리
파일 크기 가드(600줄) 위반 해소 — skill-manager.ts 630줄. 순수 함수 4개를
전용 모듈로 옮기고 기존 사용처 호환을 위해 re-export 한다.

skill-manager.ts 567줄 / skill-triggers.ts 78줄. 테스트 26건 통과.
